+++ pintos/src/threads/thread.h
@@ -88,7 +88,7 @@ struct thread
     char name[16];                      /* Name (for debugging purposes). */
     uint8_t *stack;                     /* Saved stack pointer. */
     int priority;                       /* Priority. */
-    int wakeup_ticks;                   /* Timer interrupt control. */
+    int64_t wakeup_ticks;                   /* Timer interrupt control. */
 
     /* Shared between thread.c and synch.c. */
     struct list_elem elem;              /* List element. */
